This commit changes the interface of complete, connect, and draw by removing their properties parameter.  Instead, properties are passed through cascade and flow constructors, or if you are running remotely, through the global $jobconf_properties set by jading.  This change resolves the merge issue wherein external properties would bash properties set internal to a flow.  It also makes configuring flows much more natural and similar to other Hadoop jobs: you get defaults, you side-effect them, and eventually those properties are the ones used for execution.
A cascade keeps a copy of the global $jobconf_properties and hands copies to its child flows.  Each flow may then modify its own properties in isolation.

This refactoring is a step on the way to completely eliminating $jobconf_properties and the global cascade and flow registries entirely.  See comment in lib/cascading/cascading.rb.