Put build status back now that we work on the same version of JRuby as Travis CI
This little hiccup has led to 0.0.10's tag being a bit further along than the code in the gem I released.  This would impact people trying to use 0.0.10 on JRuby 1.7.0, which was unsupported at the time.  Further, the current "fix" is only appropriate for people who package Cascading in the same jar as their application (which is likely most people).